# contracts/http/swagger.yaml
basePath: /
definitions:
...
  model.ActionExecutionStatus:
    enum:
    - pending
    - running
    - succeeded
    - failed
    - retrying
    - skipped
    - cancelled
    type: string
    x-enum-varnames:
    - ActionExecutionPending
    - ActionExecutionRunning
    - ActionExecutionSucceeded
    - ActionExecutionFailed
    - ActionExecutionRetrying
    - ActionExecutionSkipped
    - ActionExecutionCancelled
  model.ActionType:
    enum:
    - send_dm
    - timeout_user
    - kick_user
    - ban_user
    - remove_timeout
    - unban_user
    type: string
    x-enum-varnames:
    - ActionSendDM
    - ActionTimeoutUser
    - ActionKickUser
    - ActionBanUser
    - ActionRemoveTimeout
    - ActionUnbanUser
...
  model.AppealStatus:
    enum:
    - pending
    - needs_information
    - accepted
    - rejected
    - closed
    type: string
    x-enum-varnames:
    - AppealStatusPending
    - AppealStatusNeedsInformation
    - AppealStatusAccepted
    - AppealStatusRejected
    - AppealStatusClosed
...
  model.CaseSource:
    enum:
    - dashboard
    - discord
    - honeypot
    - v4_import
    type: string
    x-enum-varnames:
    - CaseSourceDashboard
    - CaseSourceDiscord
    - CaseSourceHoneypot
    - CaseSourceV4Import
  model.CaseValidity:
    enum:
    - valid
    - voided
    type: string
    x-enum-varnames:
    - CaseValidityValid
    - CaseValidityVoided
  model.ContextFieldType:
    enum:
    - short_text
    - long_text
    - boolean
    - number
    - discord_message_link
    type: string
    x-enum-varnames:
    - ContextFieldShortText
    - ContextFieldLongText
    - ContextFieldBoolean
    - ContextFieldNumber
    - ContextFieldMessageLink
...
  quack.CaseActionResponse:
    properties:
      action_type:
        $ref: '#/definitions/model.ActionType'
      id:
        type: string
      idempotency_key:
        type: string
      irreversible:
        type: boolean
      max_retries:
        type: integer
      notification_type:
        type: string
      notify_user:
        type: boolean
      position:
        type: integer
      retry_backoff_ms:
        type: integer
      safe_for_retry:
        type: boolean
      status:
        $ref: '#/definitions/model.ActionExecutionStatus'
      template_action_id:
        type: string
    type: object
  quack.CaseContextValueInput:
    properties:
      key:
        type: string
      value:
        type: object
    type: object
  quack.CaseContextValueResponse:
    properties:
      key:
        type: string
      label:
        type: string
      required:
        type: boolean
      type:
        $ref: '#/definitions/model.ContextFieldType'
      value: {}
    type: object
  quack.CaseListResponse:
    properties:
      cases:
        items:
          $ref: '#/definitions/quack.CaseResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  quack.CaseProfileResponse:
    properties:
      cases:
        items:
          $ref: '#/definitions/quack.CaseResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      summary:
        $ref: '#/definitions/quack.CaseProfileSummary'
      total:
        type: integer
    type: object
  quack.CaseProfileSummary:
    properties:
      by_template:
        additionalProperties:
          format: int64
          type: integer
        type: object
      by_validity:
        additionalProperties:
          format: int64
          type: integer
        type: object
      total:
        type: integer
    type: object
  quack.CaseResponse:
    properties:
      actions:
        items:
          $ref: '#/definitions/quack.CaseActionResponse'
        type: array
      case_number:
        type: integer
      context_channel_discord_id:
        type: string
      context_message_discord_id:
        type: string
      context_url:
        type: string
      context_values:
        items:
          $ref: '#/definitions/quack.CaseContextValueResponse'
        type: array
      created_at:
        type: string
      guild_id:
        type: string
      id:
        type: string
      metadata: {}
      moderator_discord_user_id:
        type: string
      reason:
        type: string
      replacement_case_id:
        type: string
      replaces_case_id:
        type: string
      selected_level:
        $ref: '#/definitions/quack.CaseSelectedLevel'
      source:
        $ref: '#/definitions/model.CaseSource'
      target_discord_user_id:
        type: string
      template_id:
        type: string
      template_version:
        type: integer
      updated_at:
        type: string
      validity:
        $ref: '#/definitions/model.CaseValidity'
      voided_at:
        type: string
      voided_reason:
        type: string
    type: object
  quack.CaseSelectedLevel:
    properties:
      id:
        type: string
      is_default:
        type: boolean
      matched_case_count:
        type: integer
      name:
        type: string
      notify_user:
        type: boolean
      position:
        type: integer
      trigger_case_count:
        type: integer
    type: object
...
  quack.MemberCaseListResponse:
    properties:
      cases:
        items:
          $ref: '#/definitions/quack.MemberCaseSummary'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  quack.MemberCaseSummary:
    properties:
      appeal_id:
        type: string
      appeal_status:
        $ref: '#/definitions/model.AppealStatus'
      appealable:
        type: boolean
      case_number:
        type: integer
      created_at:
        type: string
      guild_id:
        type: string
      id:
        type: string
      official_reason:
        type: string
      selected_outcome:
        $ref: '#/definitions/quack.CaseSelectedLevel'
      validity:
        $ref: '#/definitions/model.CaseValidity'
    type: object
...
  routes.caseCreateRequest:
    properties:
      context_channel_discord_id:
        type: string
      context_message_discord_id:
        type: string
      context_url:
        type: string
      context_values:
        items:
          $ref: '#/definitions/quack.CaseContextValueInput'
        type: array
      evidence_links:
        items:
          type: string
        type: array
      metadata:
        type: object
      replaces_case_id:
        type: string
      target_discord_user_id:
        type: string
      template_id:
        type: string
    type: object
...
  routes.voidCaseRequest:
    properties:
      reason:
        type: string
      replacement_case_id:
        type: string
    type: object
info:
  contact: {}
  description: HTTP boundary exposed by the Quack v5 backend.
  title: Quack HTTP API
  version: "5.0"
paths:
...
  /guilds:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: List manageable Discord guilds
      tags:
      - Guilds
...
  /guilds/{discordGuildID}/cases:
    get:
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      - description: Target Discord user ID
        in: query
        name: target_discord_user_id
        type: string
      - description: Template ID
        in: query
        name: template_id
        type: string
      - description: Case validity
        in: query
        name: validity
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quack.CaseListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: List guild cases
      tags:
      - Cases
    post:
      consumes:
      - application/json
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Retry-safe request key
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Case definition
        in: body
        name: case
        required: true
        schema:
          $ref: '#/definitions/routes.caseCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Create a moderation case
      tags:
      - Cases
  /guilds/{discordGuildID}/cases/{caseRef}:
    get:
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Case ID or number
        in: path
        name: caseRef
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get a guild case
      tags:
      - Cases
...
  /guilds/{discordGuildID}/cases/{caseRef}/void:
    post:
      consumes:
      - application/json
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Case ID or number
        in: path
        name: caseRef
        required: true
        type: string
      - description: Retry-safe request key
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Void reason and replacement
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.voidCaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Void a moderation case
      tags:
      - Cases
  /guilds/{discordGuildID}/me:
    get:
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get current guild staff context
      tags:
      - Guilds
...
  /guilds/{discordGuildID}/users/{targetDiscordUserID}/cases:
    get:
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Target Discord user ID
        in: path
        name: targetDiscordUserID
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quack.CaseProfileResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: List a guild member's case history
      tags:
      - Cases
...
  /members/me/cases/{caseID}:
    get:
      parameters:
      - description: Case ID
        in: path
        name: caseID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get the current member's case
      tags:
      - Member cases
...
  /members/me/guilds/{guildID}/cases:
    get:
      parameters:
      - description: Quack guild ID
        in: path
        name: guildID
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quack.MemberCaseListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: List the current member's cases
      tags:
      - Member cases
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    in: header
    name: Cookie
    type: apiKey
  MetricsKey:
    in: header
    name: X-Quack-Metrics-Key
    type: apiKey
  OpsKey:
    in: header
    name: X-Quack-Ops-Key
    type: apiKey
swagger: "2.0"